Update patch set 9
Patch Set 9: Code-Review-2

(3 comments)

It is not clear to me why you changed/modified the command pattern that was in place and why you change its package definition ...
I'm really against removing all the UT that was made! For which reason re you doing that?

Based on the commit message, this patch should only switch to use DTCL and add abstraction on top of it

Patch-set: 9
Label: Code-Review=-2