Update patch set 9
Patch Set 9:

(2 comments)

the AbstractCreateCommand and AbstractUpdateCommand are identical there is no clear reason why they are separated  and there is no need for AbstarctDeleteCommand as we move to the DataTreeChnageListner we always have DataTreeModification object at all cases. I replace those 3 abstract class with the AbstractCommand, so all old command classes are not good with the new abstarctCommand. I just change the package name to distinguish the new command classes we can get it back to command instead of commands.

Patch-set: 9
Label: Verified=0