Update patch set 1
Patch Set 1:

> Patch Set 1: Verified+1 Code-Review+2

What is the convention for naming new "features"? I was assuming that there are the package of YANG models, and then there is API. Case in point, legato-api and then there is features/odl-unimgr-legato-api. But in this case we seem to have merged the two.

Even so, can we call the whole thing "bgp"? 

Note, I plan to follow up on these commits with "evpn", "l2vpn", "vxlan", all IETF YANG models for each technology. So we might want to think about how we want to structure all the models.

Patch-set: 1