Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Verified+1 Code-Review+2
> 
> What is the convention for naming new "features"? I was assuming that there are the package of YANG models, and then there is API. Case in point, legato-api and then there is features/odl-unimgr-legato-api. But in this case we seem to have merged the two.
> 
> Even so, can we call the whole thing "bgp"? 
> 
> Note, I plan to follow up on these commits with "evpn", "l2vpn", "vxlan", all IETF YANG models for each technology. So we might want to think about how we want to structure all the models.

I was lazy and just added the bgp-api bundle into the odl-unimgr-rest feature. There's not really a need for a feature per module, but we do tend to have them, especially for driver modules. That lets you install individual drivers.

The legato-api is not a great example – it contains both YANG models and implementation code. It is better to separate these into, for example, legato-api and legato-impl. This makes it possible to have a dependency on the model binding APIs without also depending on implementation code.

As for the upcoming YANG models, even the bgp-api probably contains more than it should. It should really just contain the BGP YANG models and have dependencies on other packages for the YANG modules that it imports. That would mean creating project packages for bfd, l3-vlan, routing, etc. I'm not advocating that we actually do this, just that there might be reason to refactor in that direction if other projects create upstream artifacts for these YANG modules.

Patch-set: 1