Update patch set 1
Patch Set 1:

Please take a look at this change and verify whether all the functionality is as expected. There was a fair bit of manual conflict resolution and migration to the Neon API versions (mdsal binding). 

A significant amount of work was required in the Legato tests to fix compilation problems with the most recent PowerMock and Mockito library versions.

Patch-set: 1